+++ assertk/src/commonMain/kotlin/assertk/assertions/support/support.kt
@@ -18,6 +18,7 @@ private fun prefix(wrap: String): String = if (wrap.length > 0) wrap[0].toString
 private fun suffix(wrap: String): String = if (wrap.length > 1) wrap[1].toString() else ""
 
 @Suppress("ComplexMethod")
+@PublishedApi
 internal fun display(value: Any?): String {
     return when (value) {
         null -> "null"