+++ assertk/src/commonTest/kotlin/test/assertk/NameTest.kt
@@ -1,6 +1,7 @@
 package test.assertk
 
 import assertk.assertThat
+import assertk.asAssert
 import assertk.assertions.isEqualTo
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -9,18 +10,28 @@ import kotlin.test.assertFailsWith
 class NameTest {
     @Test
     fun with_no_name_fails_with_default_error_message() {
-        val error = assertFailsWith<AssertionError> {
+        val error1 = assertFailsWith<AssertionError> {
             assertThat("yes").isEqualTo("no")
         }
-        assertEquals("expected:<\"[no]\"> but was:<\"[yes]\">", error.message)
+        val error2 = assertFailsWith<AssertionError> {
+            "yes".asAssert().isEqualTo("no")
+        }
+        val expected = "expected:<\"[no]\"> but was:<\"[yes]\">"
+        assertEquals(expected, error1.message)
+        assertEquals(expected, error2.message)
     }
 
     @Test
     fun with_name_fails_with_name_prefixing_message() {
-        val error = assertFailsWith<AssertionError> {
+        val error1 = assertFailsWith<AssertionError> {
             assertThat("yes", name = "test").isEqualTo("no")
         }
-        assertEquals("expected [test]:<\"[no]\"> but was:<\"[yes]\">", error.message)
+        val error2 = assertFailsWith<AssertionError> {
+            "yes".asAssert(name = "test").isEqualTo("no")
+        }
+        val expected = "expected [test]:<\"[no]\"> but was:<\"[yes]\">"
+        assertEquals(expected, error1.message)
+        assertEquals(expected, error2.message)
     }
 
     @Test
@@ -29,10 +40,15 @@ class NameTest {
 
         val p = Person("Yoda")
 
-        val error = assertFailsWith<AssertionError> {
+        val error1 = assertFailsWith<AssertionError> {
             assertThat(p::name).isEqualTo("Darth")
         }
-        assertEquals("expected [name]:<\"[Darth]\"> but was:<\"[Yoda]\">", error.message)
+        val error2 = assertFailsWith<AssertionError> {
+            p::name.asAssert().isEqualTo("Darth")
+        }
+        val expected = "expected [name]:<\"[Darth]\"> but was:<\"[Yoda]\">"
+        assertEquals(expected, error1.message)
+        assertEquals(expected, error2.message)
     }
 
     @Test
@@ -41,9 +57,14 @@ class NameTest {
 
         val p = Person("Yoda")
 
-        val error = assertFailsWith<AssertionError> {
+        val error1 = assertFailsWith<AssertionError> {
             assertThat(p::name, "test").isEqualTo("Darth")
         }
-        assertEquals("expected [test]:<\"[Darth]\"> but was:<\"[Yoda]\">", error.message)
+        val error2 = assertFailsWith<AssertionError> {
+            p::name.asAssert("test").isEqualTo("Darth")
+        }
+        val expected = "expected [test]:<\"[Darth]\"> but was:<\"[Yoda]\">"
+        assertEquals(expected, error1.message)
+        assertEquals(expected, error2.message)
     }
 }